WIP - start refactoring functions into the new Model class
