--- packages/rev-models/src/models/index.ts
@@ -1,5 +1,0 @@
-
-export { IModel } from './model';
-export { IModelMeta } from './meta';
-export * from './operations';
-export * from './utils';
