+++ packages/rev-models/src/operations/operation.ts
@@ -1,0 +1,6 @@
+import { IWhereQuery } from '../queries/query';
+
+export interface IModelOperation {
+    name: string;
+    where?: IWhereQuery;
+}
